Support for defining new campaign regions in mods, for hota h3c
