+++ config/campaignRegions.json
@@ -1,0 +1,238 @@
+{
+	// RoE
+	
+	"good1" : {
+		"prefix": "G1",
+		"colorSuffixLength": 1,
+		"desc": [
+			{ "infix": "A", "x": 57, "y": 314 },
+			{ "infix": "B", "x": 137, "y": 309 },
+			{ "infix": "C", "x": 44, "y": 163 }
+		]
+	},
+
+	"good2" : {
+		"prefix": "G2",
+		"colorSuffixLength": 1,
+		"desc": [
+			{ "infix": "A", "x": 56, "y": 90 },
+			{ "infix": "B", "x": 316, "y": 49 },
+			{ "infix": "C", "x": 54, "y": 378 },
+			{ "infix": "D", "x": 151, "y": 126 }
+		]
+	},
+
+	"good3" : {
+		"prefix": "G3",
+		"colorSuffixLength": 1,
+		"desc": [
+			{ "infix": "A", "x": 289, "y": 376 },
+			{ "infix": "B", "x": 60, "y": 147 },
+			{ "infix": "C", "x": 131, "y": 202 }
+		]
+	},
+
+	"evil1" : {
+		"prefix": "E1",
+		"colorSuffixLength": 1,
+		"desc": [
+			{ "infix": "A", "x": 270, "y": 332 },
+			{ "infix": "B", "x": 138, "y": 113 },
+			{ "infix": "C", "x": 26, "y": 70 },
+			{ "infix": "P1", "x": 256, "y": 127 },
+			{ "infix": "P2", "x": 57, "y": 314 },
+			{ "infix": "P3", "x": 137, "y": 310 },
+			{ "infix": "P4", "x": 44, "y": 163 }
+		]
+	},
+
+	"evil2" : {
+		"prefix": "E2",
+		"colorSuffixLength": 1,
+		"desc": [
+			{ "infix": "A", "x": 131, "y": 202 },
+			{ "infix": "B", "x": 60, "y": 145 },
+			{ "infix": "C", "x": 92, "y": 261 },
+			{ "infix": "D", "x": 218, "y": 307 }
+		]
+	},
+
+	"neutral1" : {
+		"prefix": "N1",
+		"colorSuffixLength": 1,
+		"desc": [
+			{ "infix": "A", "x": 42, "y": 94 },
+			{ "infix": "B", "x": 309, "y": 290 },
+			{ "infix": "CD", "x": 188, "y": 202 }
+		]
+	},
+
+	"secret1" : {
+		"prefix": "S1",
+		"colorSuffixLength": 1,
+		"desc": [
+			{ "infix": "A", "x": 263, "y": 199 },
+			{ "infix": "B", "x": 182, "y": 210 },
+			{ "infix": "C", "x": 82, "y": 152 }
+		]
+	},
+
+	// AB
+
+	"dragonSlayer" : {
+		"prefix": "BR",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 18, "y": 233 },
+			{ "infix": "B", "x": 125, "y": 381 },
+			{ "infix": "C", "x": 224, "y": 357 },
+			{ "infix": "D", "x": 192, "y": 320 }
+		]
+	},
+
+	"foolhardyWaywardness" : {
+		"prefix": "IS",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 294, "y": 399 },
+			{ "infix": "B", "x": 183, "y": 293 },
+			{ "infix": "C", "x": 40, "y": 92 },
+			{ "infix": "D", "x": 294, "y": 398 }
+		]
+	},
+
+	"festivalOfLife" : {
+		"prefix": "KR",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 148, "y": 323 },
+			{ "infix": "B", "x": 192, "y": 235 },
+			{ "infix": "C", "x": 136, "y": 158 },
+			{ "infix": "D", "x": 87, "y": 107 }
+		]
+	},
+
+	"dragonsBlood" : {
+		"prefix": "NI",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 118, "y": 111 },
+			{ "infix": "B", "x": 223, "y": 145 },
+			{ "infix": "C", "x": 320, "y": 213 },
+			{ "infix": "D", "x": 233, "y": 250 }
+		]
+	},
+
+	"playingWithFire" : {
+		"prefix": "TA",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 228, "y": 233 },
+			{ "infix": "B", "x": 147, "y": 194 },
+			{ "infix": "C", "x": 112, "y": 97 }
+		]
+	},
+
+	"armageddonsBlade" : {
+		"prefix": "AR",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 135, "y": 238 },
+			{ "infix": "B", "x": 135, "y": 121 },
+			{ "infix": "C", "x": 206, "y": 155 },
+			{ "infix": "D", "x": 105, "y": 397 },
+			{ "infix": "E", "x": 109, "y": 275 },
+			{ "infix": "F", "x": 158, "y": 188 },
+			{ "infix": "G", "x": 200, "y": 261 },
+			{ "infix": "H", "x": 232, "y": 197 }
+		]
+	},
+
+	// SoD
+
+	"hackAndSlash" : {
+		"prefix": "HS",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 141, "y": 326 },
+			{ "infix": "B", "x": 238, "y": 275 },
+			{ "infix": "C", "x": 22, "y": 161 },
+			{ "infix": "D", "x": 5, "y": 9 }
+		]
+	},
+
+	"birthOfBarbarian" : {
+		"prefix": "BB",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 167, "y": 342 },
+			{ "infix": "B", "x": 217, "y": 263 },
+			{ "infix": "C", "x": 0, "y": 71 },
+			{ "infix": "D", "x": 291, "y": 79 },
+			{ "infix": "E", "x": 316, "y": 199 }
+		]
+	},
+
+	"newBeginning" : {
+		"prefix": "NB",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 6, "y": 292 },
+			{ "infix": "B", "x": 161, "y": 334 },
+			{ "infix": "C", "x": 63, "y": 195 },
+			{ "infix": "D", "x": 56, "y": 46 }
+		]
+	},
+
+	"elixirOfLife" : {
+		"prefix": "EL",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 11, "y": 73 },
+			{ "infix": "B", "x": 0, "y": 241 },
+			{ "infix": "C", "x": 254, "y": 34 },
+			{ "infix": "D", "x": 91, "y": 144 }
+		]
+	},
+
+	"riseOfTheNecromancer" : {
+		"prefix": "RN",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 84, "y": 319 },
+			{ "infix": "B", "x": 194, "y": 275 },
+			{ "infix": "C", "x": 67, "y": 185 },
+			{ "infix": "D", "x": 77, "y": 30 }
+		]
+	},
+
+	"unholyAlliance" : {
+		"prefix": "UA",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 157, "y": 409 },
+			{ "infix": "B", "x": 62, "y": 346 },
+			{ "infix": "C", "x": 8, "y": 8 },
+			{ "infix": "D", "x": 206, "y": 1 },
+			{ "infix": "E", "x": 132, "y": 357 },
+			{ "infix": "F", "x": 184, "y": 83 },
+			{ "infix": "G", "x": 159, "y": 263 },
+			{ "infix": "H", "x": 108, "y": 173 },
+			{ "infix": "I", "x": 55, "y": 127 },
+			{ "infix": "J", "x": 9, "y": 252 },
+			{ "infix": "K", "x": 210, "y": 176 },
+			{ "infix": "L", "x": 260, "y": 210 }
+		]
+	},
+
+	"spectreOfPower" : {
+		"prefix": "SP",
+		"colorSuffixLength": 2,
+		"desc": [
+			{ "infix": "A", "x": 7, "y": 295 },
+			{ "infix": "B", "x": 44, "y": 141 },
+			{ "infix": "C", "x": 141, "y": 21 },
+			{ "infix": "D", "x": 243, "y": 156 }
+		]
+	}
+}
