+++ lib/campaign/CampaignState.cpp
@@ -33,151 +33,6 @@ void CampaignScenario::loadPreconditionRegions(ui32 regions)
 	}
 }
 
-CampaignRegions::RegionDescription CampaignRegions::RegionDescription::fromJson(const JsonNode & node)
-{
-	CampaignRegions::RegionDescription rd;
-	rd.infix = node["infix"].String();
-	rd.pos = Point(static_cast<int>(node["x"].Float()), static_cast<int>(node["y"].Float()));
-	if(!node["labelPos"].isNull())
-		rd.labelPos = Point(static_cast<int>(node["labelPos"]["x"].Float()), static_cast<int>(node["labelPos"]["y"].Float()));
-	else
-		rd.labelPos = std::nullopt;
-	return rd;
-}
-
-JsonNode CampaignRegions::RegionDescription::toJson(CampaignRegions::RegionDescription & rd)
-{
-	JsonNode node;
-	node["infix"].String() = rd.infix;
-	node["x"].Float() = rd.pos.x;
-	node["y"].Float() = rd.pos.y;
-	if(rd.labelPos != std::nullopt)
-	{
-		node["labelPos"]["x"].Float() = (*rd.labelPos).x;
-		node["labelPos"]["y"].Float() = (*rd.labelPos).y;
-	}
-	else
-		node["labelPos"].clear();
-	return node;
-}
-
-CampaignRegions CampaignRegions::fromJson(const JsonNode & node)
-{
-	CampaignRegions cr;
-	cr.campPrefix = node["prefix"].String();
-	cr.colorSuffixLength = static_cast<int>(node["colorSuffixLength"].Float());
-	cr.campSuffix = node["suffix"].isNull() ? std::vector<std::string>() : std::vector<std::string>{node["suffix"].Vector()[0].String(), node["suffix"].Vector()[1].String(), node["suffix"].Vector()[2].String()};
-	cr.campBackground = node["background"].isNull() ? "" : node["background"].String();
-
-	for(const JsonNode & desc : node["desc"].Vector())
-		cr.regions.push_back(CampaignRegions::RegionDescription::fromJson(desc));
-
-	return cr;
-}
-
-JsonNode CampaignRegions::toJson(CampaignRegions cr)
-{
-	JsonNode node;
-	node["prefix"].String() = cr.campPrefix;
-	node["colorSuffixLength"].Float() = cr.colorSuffixLength;
-	if(cr.campSuffix.empty())
-		node["suffix"].clear();
-	else
-		node["suffix"].Vector() = JsonVector{ JsonNode(cr.campSuffix[0]), JsonNode(cr.campSuffix[1]), JsonNode(cr.campSuffix[2]) };
-	if(cr.campBackground.empty())
-		node["background"].clear();
-	else
-		node["background"].String() = cr.campBackground;
-	node["desc"].Vector() = JsonVector();
-	for(auto & region : cr.regions)
-		node["desc"].Vector().push_back(CampaignRegions::RegionDescription::toJson(region));
-	return node;
-}
-
-CampaignRegions CampaignRegions::getLegacy(int campId)
-{
-	static std::vector<CampaignRegions> campDescriptions;
-	if(campDescriptions.empty()) //read once
-	{
-		const JsonNode config(JsonPath::builtin("config/campaign_regions.json"));
-		for(const JsonNode & campaign : config["campaign_regions"].Vector())
-			campDescriptions.push_back(CampaignRegions::fromJson(campaign));
-	}
-
-	return campDescriptions.at(campId);
-}
-
-ImagePath CampaignRegions::getBackgroundName() const
-{
-	if(campBackground.empty())
-		return ImagePath::builtin(campPrefix + "_BG.BMP");
-	else
-		return ImagePath::builtin(campBackground);
-}
-
-Point CampaignRegions::getPosition(CampaignScenarioID which) const
-{
-	auto const & region = regions[which.getNum()];
-	return region.pos;
-}
-
-std::optional<Point> CampaignRegions::getLabelPosition(CampaignScenarioID which) const
-{
-	auto const & region = regions[which.getNum()];
-	return region.labelPos;
-}
-
-ImagePath CampaignRegions::getNameFor(CampaignScenarioID which, int colorIndex, const std::string & type) const
-{
-	auto const & region = regions[which.getNum()];
-
-	static const std::array<std::array<std::string, 8>, 3> colors = {{
-		{ "", "", "", "", "", "", "", "" },
-		{ "R", "B", "N", "G", "O", "V", "T", "P" },
-		{ "Re", "Bl", "Br", "Gr", "Or", "Vi", "Te", "Pi" }
-	}};
-
-	std::string color = colors[colorSuffixLength][colorIndex];
-
-	return ImagePath::builtin(campPrefix + region.infix + "_" + type + color + ".BMP");
-}
-
-ImagePath CampaignRegions::getAvailableName(CampaignScenarioID which, int color) const
-{
-	if(campSuffix.empty())
-		return getNameFor(which, color, "En");
-	else
-		return getNameFor(which, color, campSuffix[0]);
-}
-
-ImagePath CampaignRegions::getSelectedName(CampaignScenarioID which, int color) const
-{
-	if(campSuffix.empty())
-		return getNameFor(which, color, "Se");
-	else
-		return getNameFor(which, color, campSuffix[1]);
-}
-
-ImagePath CampaignRegions::getConqueredName(CampaignScenarioID which, int color) const
-{
-	if(campSuffix.empty())
-		return getNameFor(which, color, "Co");
-	else
-		return getNameFor(which, color, campSuffix[2]);
-}
-
-void CampaignHeader::loadLegacyData(ui8 campId)
-{
-	campaignRegions = CampaignRegions::getLegacy(campId);
-	numberOfScenarios = LIBRARY->generaltexth->getCampaignLength(campId);
-}
-
-void CampaignHeader::loadLegacyData(const CampaignRegions & regions, int numOfScenario)
-{
-	campaignRegions = regions;
-	numberOfScenarios = numOfScenario;
-}
-
 bool CampaignHeader::playerSelectedDifficulty() const
 {
 	return difficultyChosenByPlayer;
@@ -526,8 +381,10 @@ void Campaign::overrideCampaign()
 	for (auto & entry : node.Struct())
 		if(filename == entry.first)
 		{
-			if(!entry.second["regions"].isNull() && !entry.second["scenarioCount"].isNull())
-				loadLegacyData(CampaignRegions::fromJson(entry.second["regions"]), entry.second["scenarioCount"].Integer());
+			if(!entry.second["regions"].isNull())
+				campaignRegions = CampaignRegions(entry.second["regions"]);
+			if (!entry.second["scenarioCount"].isNull())
+				numberOfScenarios = entry.second["scenarioCount"].Integer();
 			if(!entry.second["loadingBackground"].isNull())
 				loadingBackground = ImagePath::builtin(entry.second["loadingBackground"].String());
 			if(!entry.second["videoRim"].isNull())
