+++ lib/texts/CGeneralTextHandler.h
@@ -42,9 +42,6 @@ class DLL_LINKAGE CGeneralTextHandler: public TextLocalizationContainer
 {
 	void readToVector(const std::string & sourceID, const std::string & sourceName);
 
-	/// number of scenarios in specific campaign. TODO: move to a better location
-	std::vector<size_t> scenariosCountPerCampaign;
-
 public:
 	LegacyTextContainer allTexts;
 
@@ -77,8 +74,6 @@ public:
 	std::vector<std::string> findStringsWithPrefix(const std::string & prefix);
 
 	int32_t pluralText(int32_t textIndex, int32_t count) const;
-
-	size_t getCampaignLength(size_t campaignID) const;
 
 	CGeneralTextHandler();
 	CGeneralTextHandler(const CGeneralTextHandler&) = delete;
